progress
Two file opening methods are supported, this is simply for studying the
behavioural differences that one can get from the two.

A new strategy is embarked on for programe structure. Light inspiration
taken from io_uring.
The user facing functions barely has any code written in them and are
instead comprised of locally written static functions.
Even if the functions are declared inline, it feels like this would be a
much simpler way to review this.

Completed right now, in limited capacity, is the ability to create a
basic template, load a recordings file and test the header for validity.

immediate goals include completing the loading and verification system
for both FILE and fileno loads and the conversion to/from CSV